move thread related events under the same package
